server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
psutil>=5.9.0
requests>=2.31.0
duckduckgo_search>=6.1.0
cryptography>=42.0.0
llama-cpp-python>=0.2.0
